requirements.txt
Django==4.2.7
kubernetes==29.0.0
PyYAML==6.0.1 